modified so it will be more clear

+++ Question_3/targil3.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 #include <math.h>
 
-int factorial (int n){ //classing function factorial rec
+long factorial (int n){ //classing function factorial rec
     if (n == 0){
         return 1;
     }
@@ -12,6 +12,6 @@ int factorial (int n){ //classing function factorial rec
 
 long double poi(int k, int delta){
     double top = pow(delta, k) * expf(-delta); // the second expression give me a double 
-    int bottom = factorial(k);
+    long bottom = factorial(k);
     return top/bottom;
 }
